requirements.txt
torch
torcheval
torchvision
pytorch-ignite
omegaconf
numpy
pandas
rich
opencv-python
pytest
matplotlib
hydra-core
scikit-learn
umap-learn
seaborn